#29 Add Postgres and MySQL support and refactor memory package
Added Postgres and MySQL datasource configurations and dependencies. Introduced a new `DemoFailProjector` for handling failure scenarios. Renamed multiple classes and packages from `query` to `memory.query` for better organization and clarity.
