+++ evento-demo/evento-demo-observer/src/main/java/com/evento/demo/memory/config/EventoConfiguration.java
@@ -1,13 +1,10 @@
-package com.evento.demo.config;
+package com.evento.demo.memory.config;
 
 import com.evento.application.EventoBundle;
 import com.evento.application.bus.ClusterNodeAddress;
 import com.evento.application.bus.EventoServerMessageBusConfiguration;
-import com.evento.common.messaging.consumer.impl.InMemoryConsumerStateStore;
 import com.evento.common.performance.ThreadCountAutoscalingProtocol;
-import com.evento.common.utils.Context;
-import com.evento.demo.DemoQueryApplication;
-import com.evento.demo.query.DemoProjector;
+import com.evento.demo.DemoObserverApplication;
 import com.evento.demo.telemetry.SentryTracingAgent;
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -15,9 +12,6 @@ import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Scope;
-
-import java.util.Map;
-import java.util.Set;
 
 @Configuration
 public class EventoConfiguration {
@@ -33,16 +27,11 @@ public class EventoConfiguration {
 			@Value("${evento.cluster.autoscaling.max.overflow}") int maxOverflow,
 			@Value("${evento.cluster.autoscaling.min.threads}") int minThreads,
 			@Value("${evento.cluster.autoscaling.max.underflow}") int maxUnderflow,
-			BeanFactory factory,
-			@Value("${spring.datasource.url}") String connectionUrl,
-			@Value("${spring.datasource.username}") String username,
-			@Value("${spring.datasource.password}") String password,
-			@Value("${sentry.dns}") String sentryDns
+			@Value("${sentry.dns}") String sentryDns,
+			BeanFactory factory
 	) throws Exception {
 		return EventoBundle.Builder.builder()
-				.setBasePackage(DemoQueryApplication.class.getPackage())
-				.setConsumerStateStoreBuilder(InMemoryConsumerStateStore::new)
-				.setInjector(factory::getBean)
+				.setBasePackage(DemoObserverApplication.class.getPackage())
 				.setBundleId(bundleId)
 				.setBundleVersion(bundleVersion)
 				.setEventoServerMessageBusConfiguration(new EventoServerMessageBusConfiguration(
@@ -58,8 +47,6 @@ public class EventoConfiguration {
 						maxOverflow,
 						maxUnderflow, 60 * 1000))
 				.setInjector(factory::getBean)
-				// .setContexts(Map.of(DemoProjector.class.getSimpleName(), Set.of(Context.DEFAULT, "other")))
 				.start();
-
 	}
 }
