+++ evento-demo/evento-demo-query/build.gradle
@@ -47,8 +47,12 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.integration:spring-integration-jdbc'
     runtimeOnly 'com.h2database:h2'
+    runtimeOnly 'org.postgresql:postgresql'
+    runtimeOnly 'com.mysql:mysql-connector-j'
 
     implementation project(":evento-demo:evento-demo-api")
+    implementation(project(':evento-consumer-state-store:evento-consumer-state-store-postgres'))
+    implementation(project(':evento-consumer-state-store:evento-consumer-state-store-mysql'))
 }
 
 tasks.register('prepareKotlinBuildScriptModel') {
