+++ evento-demo/evento-demo-query/src/main/resources/application.properties
@@ -21,3 +21,11 @@ spring.h2.console.enabled=true
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.ddl-auto=create
 spring.jpa.properties.hibernate.globally_quoted_identifiers=true
+
+spring.postgres.datasource.password=secret
+spring.postgres.datasource.url=jdbc:postgresql://localhost:5432/evento-demo
+spring.postgres.datasource.username=postgres
+
+spring.mysql.datasource.password=secret
+spring.mysql.datasource.url=jdbc:mysql://localhost:3306/evento-demo
+spring.mysql.datasource.username=root
